refactor(ui): replace Titles component with SubMenu for improved functionality
The Titles component was replaced with the new SubMenu component across multiple files to enhance navigation and provide additional features like quick access to the dashboard and a back button. The old Titles component was deleted as it is no longer needed
