+++ src/features/information/components/Information.tsx
@@ -1,5 +1,6 @@
 import { useParams } from "react-router";
 import InfoPage from "./InfoPage";
+import SubMenu from "../../../components/ui/SubMenu";
 
 
 function Information() {
@@ -7,6 +8,7 @@ function Information() {
 
     return (
         <div className="flex flex-col gap-2 w-full h-full">
+            <SubMenu title="Información destacada" />
             <InfoPage id={id!} />
         </div>
     );
